ARM: 6674/1: LPAE: use long long format when printing physical addresses and ptes
For the Kernel to support 2 level and 3 level page tables, physical
addresses (and also page table entries) need to be 32 or 64-bits depending
upon the configuration.

This patch uses the %08llx conversion specifier for physical addresses
and page table entries, ensuring that they are cast to (long long) so
that common code can be used regardless of the datatype widths.

Acked-by: Catalin Marinas <catalin.marinas@arm.com>

This patch was imported from linux-omap git commit id
29a38193c15bd72ba96c57a805443ef46ea6a6ac. Since it was on latest kernel (3.x),
removed the files which were not needed to be patched in 2.6.37 to avoid
warnings.

In addition, removed similar warning form plat-omap/sram.c due to usage of
phys_addr_t in __pfn_to_phys().

+++ arch/arm/plat-omap/sram.c
@@ -222,8 +222,8 @@ static void __init omap_map_sram(void)
 	iotable_init(omap_sram_io_desc, ARRAY_SIZE(omap_sram_io_desc));
 
 	printk(KERN_INFO "SRAM: Mapped pa 0x%08lx to va 0x%08lx size: 0x%lx\n",
-	__pfn_to_phys(omap_sram_io_desc[0].pfn),
-	omap_sram_io_desc[0].virtual,
+	       (long unsigned int)__pfn_to_phys(omap_sram_io_desc[0].pfn),
+	       omap_sram_io_desc[0].virtual,
 	       omap_sram_io_desc[0].length);
 
 	/*
